traffic_prediction/requirements.txt
streamlit
numpy
pandas
torch_geometric
torch-geometric-temporal
opencv-python
ultralytics
pillow